refactor(ui): centralize character tool catalog for picker and wizard
